Datasets: Split Dataset and PublicDataset into own modules.
