+++ src/pymovements/datasets/__init__.py
@@ -18,7 +18,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 """This module provides access to publicly available datasets."""
-from pymovements.datasets.base import Dataset  # noqa: F401
-from pymovements.datasets.base import PublicDataset  # noqa: F401
+from pymovements.datasets.dataset import Dataset  # noqa: F401
 from pymovements.datasets.gazebase import GazeBase  # noqa: F401
 from pymovements.datasets.judo1000 import JuDo1000  # noqa: F401
+from pymovements.datasets.public_dataset import PublicDataset  # noqa: F401
